+++ .gitignore
@@ -23,4 +23,4 @@ __pycache__
 
 # data folder to ignore
 data
-polus/manifests
+polus/_plugins/manifests/*
