+++ polus/_plugins/classes/__init__.py
@@ -1,0 +1,3 @@
+from .plugin_classes import Plugin, ComputePlugin, load_plugin, submit_plugin, _Plugins
+
+__all__ = ["Plugin", "ComputePlugin", "load_plugin", "submit_plugin", "_Plugins"]
