--- polus/gh/__init__.py
@@ -1,3 +1,0 @@
-from .gh import add_plugin_from_gh, _init_github
-
-__all__ = ["add_plugin_from_gh", "_init_github"]
