--- polus/io/__init__.py
@@ -1,19 +1,0 @@
-from .io import (
-    IOBase,
-    Version,
-    Input,
-    Output,
-    DuplicateVersionFound,
-    _in_old_to_new,
-    _ui_old_to_new,
-)
-
-__all__ = [
-    "IOBase",
-    "Version",
-    "Input",
-    "Output",
-    "DuplicateVersionFound",
-    "_in_old_to_new",
-    "_ui_old_to_new",
-]
