--- polus/manifest_utils/__init__.py
@@ -1,15 +1,0 @@
-from .manifest_utils import (
-    is_valid_manifest,
-    _load_manifest,
-    _scrape_manifests,
-    validate_manifest,
-    _error_log,
-)
-
-__all__ = [
-    "is_valid_manifest",
-    "_load_manifest",
-    "_scrape_manifests",
-    "validate_manifest",
-    "_error_log",
-]
