--- polus/plugin_classes/__init__.py
@@ -1,3 +1,0 @@
-from .plugins import Plugin, ComputePlugin, load_plugin, submit_plugin, _Plugins
-
-__all__ = ["Plugin", "ComputePlugin", "load_plugin", "submit_plugin", "_Plugins"]
