--- polus/registry/__init__.py
@@ -1,3 +1,0 @@
-from .registry import WippPluginRegistry
-
-__all__ = ["WippPluginRegistry"]
