--- polus/utils/__init__.py
@@ -1,3 +1,0 @@
-from .utils import name_cleaner, cast_version
-
-__all__ = ["name_cleaner", "cast_version"]
