+++ setup.py
@@ -8,7 +8,7 @@ with open("./polus/_plugins/VERSION", "r") as fh:
     with open("./polus/_plugins/VERSION", "w") as fw:
         fw.write(version)
 
-package_data = ["_plugins/VERSION", "manifests/*"]
+package_data = ["_plugins/VERSION", "_plugins/manifests/*"]
 
 setup(
     name="polus-plugins",
